fix: Fix time zone update
With a large number of cities on the map (I had an error with 7-8
cities) time zone for them does not have time to update, due to
limitations api service timezonedb.com

And, although the documentation
of the service specifies the limit of not more than one request per
second, in fact, about half of the cities shows the wrong time.

By
increasing the delay between updates from 1100 to 3000 milliseconds,
this problem disappears.

Translated with www.DeepL.com/Translator (free
version)